Add files via upload
1st semester code examined, issues found and corrected. 
Data visualisation : We did not select just the features we wanted to visualise, Corrected.
Data cleaning: There was one lowercase 'f' in the gender column - now changed to 'F'
Data Splitting vs. Encoding Order: We split the data into training and testing sets but then used one-hot encode on the whole df. This order was incorrect.
Treatment of the Gender column: We used onehot encoder to code the gender column but later in the code we also used pd.getdummies to do the same thing again. 
StandardScaler- We ran scaler = StandardScaler() twice but only used it once.